+++ packages/heap/src/Error/HeapError.ts
@@ -1,7 +1,9 @@
-import { DataSourceError } from '@jamashita/anden-error';
+import { DataSourceError } from '@jamashita/catacombe-datasource';
 
-export class HeapError extends DataSourceError<'HeapError', 'Heap'> {
+export class HeapError extends DataSourceError<'HeapError'> {
+  public readonly noun: 'HeapError' = 'HeapError';
+
   public constructor(message: string, cause?: Error) {
-    super('HeapError', 'Heap', message, cause);
+    super('HeapError', message, cause);
   }
 }