+++ packages/mysql/src/Error/MySQLError.ts
@@ -1,7 +1,9 @@
-import { DataSourceError } from '@jamashita/anden-error';
+import { DataSourceError } from '@jamashita/catacombe-datasource';
 
-export class MySQLError extends DataSourceError<'MySQLError', 'MySQL'> {
+export class MySQLError extends DataSourceError<'MySQLError'> {
+  public readonly noun: 'MySQLError' = 'MySQLError';
+
   public constructor(message: string, cause?: Error) {
-    super('MySQLError', 'MySQL', message, cause);
+    super('MySQLError', message, cause);
   }
 }