+++ packages/request/src/Error/RequestError.ts
@@ -1,7 +1,9 @@
-import { DataSourceError } from '@jamashita/anden-error';
+import { DataSourceError } from '@jamashita/catacombe-datasource';
 
-export class RequestError extends DataSourceError<'RequestError', 'Request'> {
+export class RequestError extends DataSourceError<'RequestError'> {
+  public readonly noun: 'RequestError' = 'RequestError';
+
   public constructor(message: string, cause?: Error) {
-    super('RequestError', 'Request', message, cause);
+    super('RequestError', message, cause);
   }
 }